Program now functions on Linux

+++ engine/render/physics.cc
@@ -94,9 +94,9 @@ LoadFromIndexBuffer(fx::gltf::Document const& doc, ColliderMesh* mesh)
     mesh->bSphereRadius = vbAccessor.max[0];
     mesh->bSphereRadius = std::max(mesh->bSphereRadius, vbAccessor.max[1]);
     mesh->bSphereRadius = std::max(mesh->bSphereRadius, vbAccessor.max[2]);
-    mesh->bSphereRadius = std::max(mesh->bSphereRadius, fabs(vbAccessor.min[0]));
-    mesh->bSphereRadius = std::max(mesh->bSphereRadius, fabs(vbAccessor.min[1]));
-    mesh->bSphereRadius = std::max(mesh->bSphereRadius, fabs(vbAccessor.min[2]));
+    mesh->bSphereRadius = std::max(mesh->bSphereRadius, (float)fabs(vbAccessor.min[0]));
+    mesh->bSphereRadius = std::max(mesh->bSphereRadius, (float)fabs(vbAccessor.min[1]));
+    mesh->bSphereRadius = std::max(mesh->bSphereRadius, (float)fabs(vbAccessor.min[2]));
 }
